use generalised `ServiceLink` component instead of individual components per service
